Moved functions from class MoleculeUtils to class StereoUtils
