+++ jna-rinchi-api/src/main/java/io/github/dan2097/jnarinchi/cheminfo/MoleculeUtils.java
@@ -32,61 +32,8 @@ import io.github.dan2097.jnarinchi.RinchiInputComponent;
 
 public class MoleculeUtils {
 	
-	public static enum MolCoordinatesType {
-		_0D, _2D, _3D
-	}
 	
-	public static int getNumberOfTetrahedralChiralityAtoms(InchiInput inchiInput) {
-		int nTH = 0;
-		for (int i = 0; i < inchiInput.getStereos().size(); i++) {
-			InchiStereo stereo = inchiInput.getStereos().get(i);
-			if (stereo.getType() == InchiStereoType.Tetrahedral)
-				if (stereo.getParity() == InchiStereoParity.ODD || 
-					stereo.getParity() == InchiStereoParity.EVEN)
-					nTH++;
-		}
-		return nTH;
-	}
 	
-	public static Map<InchiAtom,InchiStereoParity> getAtomParities(InchiInput inchiInput) {
-		Map<InchiAtom,InchiStereoParity> parities = new HashMap<>();
-		for (int i = 0; i < inchiInput.getStereos().size(); i++) {
-			InchiStereo stereo = inchiInput.getStereos().get(i);
-			if (stereo.getType() == InchiStereoType.Tetrahedral)
-				if (stereo.getParity() == InchiStereoParity.ODD || 
-					stereo.getParity() == InchiStereoParity.EVEN || 
-					stereo.getParity() == InchiStereoParity.UNKNOWN) {
-					parities.put(stereo.getCentralAtom(), stereo.getParity());
-				}	
-		}
-		return parities;
-	}
-	
-	public static InchiStereo createTetrahedralStereo(RinchiInputComponent ric, InchiAtom atom, InchiStereoParity parity) {
-		List<InchiAtom> neighbAtoms = ric.getConectedAtomList(atom);
-		if (neighbAtoms.size() < 3 || neighbAtoms.size() > 4)
-			return null; //Unable to create stereo element
-		
-		if (neighbAtoms.size() == 3) { 
-			if (atom.getImplicitHydrogen() == 1) {
-				if (containsHydrogen(neighbAtoms))
-					return null; //one implicit and one explicit hydogen neighbors
-				
-				neighbAtoms.add(InchiStereo.STEREO_IMPLICIT_H);
-			}	
-			else {
-				//TODO check for lone pair
-				return null; //Unable to create stereo element
-			}	
-		}
-		
-		InchiAtom[] sortedAtoms = sortAtomsToBeWithIncreasingIndices(ric, neighbAtoms);
-		
-		InchiStereo stereo = InchiStereo.createTetrahedralStereo(atom, 
-				sortedAtoms[0], sortedAtoms[1], sortedAtoms[2], sortedAtoms[3], parity);
-		
-		return stereo;
-	}
 	
 	public static boolean containsHydrogen(List<InchiAtom> atoms) {
 		for (InchiAtom a: atoms)
@@ -95,39 +42,6 @@ public class MoleculeUtils {
 		return false;
 	}
 	
-	public static InchiAtom[] sortAtomsToBeWithIncreasingIndices(InchiInput inchiInput, List<InchiAtom> atoms) {
-		if (atoms == null)
-			return null;
-		InchiAtom[] sorted = atoms.toArray(new InchiAtom[] {});
-		int n = atoms.size();
-		
-		if (n <= 1)
-			return sorted;
-		
-		if (n == 2) {
-			if (inchiInput.getAtoms().indexOf(sorted[0]) > inchiInput.getAtoms().indexOf(sorted[1]))
-				swap(0, 1, sorted);
-			return sorted;
-		}
-		
-		//get atom indices
-		Map<InchiAtom, Integer> atomIndices = new HashMap<>();
-		for (int i = 0; i < n; i++)
-			atomIndices.put(sorted[i], inchiInput.getAtoms().indexOf(sorted[i]));
-		//bubble sorting
-		for (int i = n-1; i >= 0; i--) 
-			for (int j = 0; j < i; j++) {
-				if (atomIndices.get(sorted[i]) > atomIndices.get(sorted[j])) 
-					swap (i,j, sorted);
-			}
-		return sorted;
-	}
-	
-	private static void swap(int i, int j, Object[] objects) {
-		Object obj = objects[i];
-		objects[i] = objects[j];
-		objects[j] = obj;
-	}
 	
 	public static void setImplicitHydrogenAtoms(InchiInput inchiInput) {
 		Map<InchiAtom,Integer> atomExplVal =  getExplicitAtomValencies(inchiInput);
@@ -177,33 +91,7 @@ public class MoleculeUtils {
 		return 0;
 	}
 	
-	public static int[] coordinateStatistics (InchiInput inchiInput) {
-		if (inchiInput == null)
-			return null;
-		
-		int nX = 0;
-		int nY = 0;
-		int nZ = 0;
-		
-		for (int i = 0; i < inchiInput.getAtoms().size(); i++) {
-			InchiAtom at = inchiInput.getAtom(i);
-			if (at.getX() != 0.0)
-				nX++;
-			if (at.getY() != 0.0)
-				nY++;
-			if (at.getZ() != 0.0)
-				nZ++;
-		}
-		return new int[] {nX, nY, nZ};
-	}
 	
-	public static MolCoordinatesType getMolCoordinatesType(int nX, int nY, int nZ) {
-		if (nZ != 0)
-			return MolCoordinatesType._3D;
-		if (nY != 0)
-			return MolCoordinatesType._2D;
-		return MolCoordinatesType._0D;
-	}
 	
 	/**
 	 * Function returns the implicit valence (h atom count) for a given atom,
